fix all test code position
