+++ jest.config.js
@@ -1,0 +1,29 @@
+const { defaults } = require('jest-config');
+
+module.exports = {
+  preset: 'ts-jest',
+  globals: {
+    'ts-jest': {
+      tsConfig: 'tsconfig.test.json',
+      diagnostics: {
+        pathRegex: /\.(spec|test)\.ts?(x)$/,
+        warnOnly: true,
+      },
+      babelConfig: true,
+    },
+  },
+  collectCoverageFrom: ['**/*.{js,jsx,ts,tsx}', '!**/*.d.ts', '!**/node_modules/**'],
+  transform: {
+    '^.+\\.(ts|tsx)?$': 'ts-jest',
+  },
+  transformIgnorePatterns: ['/node_modules/', '^.+\\.module\\.(css|sass|scss)$'],
+  testPathIgnorePatterns: ['/node_modules/', '/.next/'],
+  testMatch: ['<rootDir>/__test__/**/*.(test|spec).ts?(x)'],
+  moduleFileExtensions: [...defaults.moduleFileExtensions, 'ts', 'tsx', 'js', 'jsx'],
+  moduleNameMapper: {
+    '^.+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2|svg)$': 'identity-obj-proxy',
+    '^.+\\.module\\.(css|sass|scss)$': 'identity-obj-proxy',
+    '^.+\\.worker.(js|ts)$': '<rootDir>/__test__/__mocks__/workerMock.ts',
+    'worker-loader?(.*)': '<rootDir>/__test__/__mocks__/workerMock.ts',
+  },
+};
