+++ packages/docs/tsconfig.json
@@ -12,8 +12,7 @@
   "include": ["./src/**/*"],
   "exclude": [
     "node_modules",
-    "src/__test__/**",
-    "src/**/*.test.ts"
+    "__test__/**",
   ],
   "typeAcquisition": {
     "enable": true
