+++ packages/localize-components-atomic/tsconfig.json
@@ -8,11 +8,10 @@
       "node"
     ],
   },
-  "include": ["./src/**/*"],
+  "include": ["./src/**/*", "__test__"],
   "exclude": [
     "node_modules",
-    "src/__test__/**",
-    "src/**/*.test.ts"
+    "__test__/**",
   ],
   "typeAcquisition": {
     "enable": true
