+++ tsconfig.json
@@ -48,7 +48,6 @@
   "exclude": [
     ".vscode",
     "node_modules/**",
-    "./src/__test__/**",
-    "./src/**/*.test.ts",
+    "__test__/**",
   ]
 }
